Corrected parseInt to parseFloat

+++ practicetools/albertadrugprice/functions.js
@@ -1285,14 +1285,14 @@ function chooseResult(selection) {
  ****************************************************************************/
 function addLCA(result) {
 	var lcaIndex = 0;
-	var lcaCost = parseInt(result[0].unit_price);
+	var lcaCost = parseFloat(result[0].unit_price);
 	var lcaEntry;
 	
 	// Identifies the index of the LCA
 	for (var i = 0; i < result.length; i++) {
-		if (parseInt(result[i].unit_price) < lcaCost) {
+		if (parseFloat(result[i].unit_price) < lcaCost) {
 			lcaIndex = i;
-			lcaCost = parseInt(result[i].unit_price);
+			lcaCost = parseFloat(result[i].unit_price);
 		}
 	}
